Move OperatorUtilsTest to UnboundedStreamIterationITCase
